refactor(docs): split playground builder output and use svelte:window popstate
Move Builder rebuild/codegen into playground-output-builder with a types leaf
and facade re-exports. Extract pipelineErrorToPlaygroundDiagnostic for
preview. Register popstate via svelte:window on Playground and the gallery.
